// apps/pc/src/generated/models/Pl_partnersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pl_partnerspl_sharesetupstatuscode = {
  100000000: '初期設定中',
  100000001: '設定完了'
} as const;
export type Pl_partnerspl_sharesetupstatuscode = keyof typeof Pl_partnerspl_sharesetupstatuscode;
export const Pl_partnerspl_tradingstatuscode = {
  100000000: '未承認',
  100000001: '取引中',
  100000002: '休止中',
  100000003: '終了'
} as const;
export type Pl_partnerspl_tradingstatuscode = keyof typeof Pl_partnerspl_tradingstatuscode;
export const Pl_partnersstatecode = {
  0: 'アクティブ',
  1: '非アクティブ'
} as const;
export type Pl_partnersstatecode = keyof typeof Pl_partnersstatecode;
export const Pl_partnersstatuscode = {
  1: 'アクティブ',
  2: '非アクティブ'
} as const;
export type Pl_partnersstatuscode = keyof typeof Pl_partnersstatuscode;

export interface Pl_partnersBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pl_aclversion: number;
  pl_address?: string;
  pl_industry?: string;
  "pl_MainOwnerLookup@odata.bind"?: string;
  pl_name?: string;
  pl_normalizedname?: string;
  pl_partnerid: string;
  pl_phone?: string;
  pl_registeredat?: string;
  "pl_RegisteredByLookup@odata.bind"?: string;
  pl_registrationcontenthash: string;
  pl_registrationkey: string;
  pl_sharesetupstatuscode: Pl_partnerspl_sharesetupstatuscode;
  pl_tradingstatuscode?: Pl_partnerspl_tradingstatuscode;
  statecode: Pl_partnersstatecode;
  statuscode?: Pl_partnersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pl_partners extends Pl_partnersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pl_mainownerlookupname?: string;
  pl_mainownerlookupyominame?: string;
  pl_registeredbylookupname?: string;
  pl_registeredbylookupyominame?: string;
  pl_sharesetupstatuscodename?: string;
  pl_tradingstatuscodename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pl_mainownerlookup?: object;
  _pl_mainownerlookup_value?: string;
  pl_registeredbylookup?: object;
  _pl_registeredbylookup_value?: string;
}